# gpt-action-openapi.yaml
openapi: 3.1.0
info:
  title: Potato Gateway Actions Safe
  description: >
    API for checking Hong's Potato multi-agent system, inspecting registered
    Hermes Agent profiles, running visible multi-agent video workflows, and
    managing structured calibration sessions.
  version: 0.2.6
  x-potato-schema-build: prompt-closed-loop-3.1-20260809
servers:
  - url: https://zhanghongmac-mini.tail282e0b.ts.net
paths:
...
  /api/calibrations/{session_id}/prompt-candidates:
    post:
      operationId: generatePromptCandidate
      summary: Generate a draft Prompt candidate from calibration evidence
      description: >
        Combines the session goal, acceptance criteria, user feedback, and
        completed critic reports into a managed addendum on the current active
        Prompt. The full Prompt remains private and the draft is not activated.
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePromptCandidateRequest"
      responses:
        "201":
          description: Draft candidate generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptVersionSummary"
        "200":
          description: Existing idempotent candidate returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptVersionSummary"
...
  /api/agents/{agent_id}/prompt-versions:
    post:
      operationId: createPromptCandidate
      summary: Store a candidate Prompt version without activating it
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            enum: [researcher, creator, critic]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptCandidateRequest"
      responses:
        "201":
          description: Candidate created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptVersionSummary"
    get:
      operationId: listPromptVersions
      summary: List Prompt version metadata without Prompt content
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            enum: [researcher, creator, critic]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Prompt versions newest first
          content:
            application/json:
              schema:
                type: object
                required: [agent_id, versions]
                properties:
                  agent_id:
                    type: string
                  versions:
                    type: array
                    items:
                      $ref: "#/components/schemas/PromptVersionSummary"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
  schemas:
...
    CreatePromptCandidateRequest:
      type: object
      additionalProperties: false
      required: [client_request_id, content, change_summary]
      properties:
        client_request_id:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9._-]+$"
        content:
          type: string
          minLength: 1
          maxLength: 200000
        change_summary:
          type: string
          minLength: 1
          maxLength: 4000
        calibration_session_id:
          type: string
          maxLength: 128
    GeneratePromptCandidateRequest:
      type: object
      additionalProperties: false
      required: [client_request_id]
      properties:
        client_request_id:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9._-]+$"
        additional_guidance:
          type: string
          maxLength: 12000
    PromptVersionSummary:
      type: object
      additionalProperties: false
      required:
        - prompt_version_id
        - agent_id
        - status
        - content_sha256
        - base_content_sha256
        - change_summary
        - managed_addendum
        - calibration_session_id
        - created_at
        - updated_at
        - activated_at
      properties:
        prompt_version_id:
          type: string
        agent_id:
          type: string
          enum: [researcher, creator, critic]
        status:
          type: string
          enum: [draft, testing, active, retired]
        content_sha256:
          type: string
          pattern: "^[0-9a-f]{64}$"
        base_content_sha256:
          type: string
        change_summary:
          type: string
        managed_addendum:
          type: string
        calibration_session_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        activated_at:
          type: string
          format: date-time
...